Fix session id

+++ platforms/android/aps-sdk/airplay/src/main/cpp/IAirPlayMirrorHandler.cpp
@@ -69,9 +69,9 @@ void IAirPlayMirrorHandler::on_mirror_stream_data(
       jbyteArray byte_array = env->NewByteArray(p->payload_size);
       env->SetByteArrayRegion(byte_array, 0, p->payload_size,
                               (jbyte *)(p->payload));
-      jlong timestamp = (uint32_t)normalize_ntp_to_ms(p->timestamp);
+      jlong timestamp = (jlong)(uint32_t)normalize_ntp_to_ms(p->timestamp);
       // convert the value
-      env->CallVoidMethod(jvm_obj_, mid, byte_array, (jlong)timestamp);
+      env->CallVoidMethod(jvm_obj_, mid, byte_array, timestamp);
       env->DeleteLocalRef(byte_array);
     } else {
       __android_log_write(ANDROID_LOG_ERROR, LOG_TAG,
